framework/requirements.txt
click==6.6
click-plugins==1.0.3
cligj==0.4.0
Django==1.7.5
django-allauth==0.19.1
django-cronjobs==0.2.3
django-filter==0.9.2
django-geojson==2.9.0
django-rest-auth==0.3.4
django-suit==0.2.12
djangorestframework==3.0.5
Fiona==1.6.3.post1
geojson==1.0.9
google-api-python-client==1.5.0
httplib2==0.9.2
lxml==3.4.2
numpy==1.10.4
oauth2client==2.0.2
oauthlib==0.7.2
OWSLib==0.8.13
pandas==0.16.0
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pynoramio==0.1.1
pysqlite==2.8.2
python-dateutil==2.4.0
python-openid==2.2.5
pytz==2014.10
requests==2.3.0
requests-oauthlib==0.4.2
rsa==3.4.2
Shapely==1.5.15
simplejson==3.8.2
six==1.9.0
uritemplate==0.6
wheel==0.24.0
